Fixed issue with Madgwick. Data was not being normalised before being passed to Madgwick also gyroscope data needed to be in radians per second.

+++ ExpressiveTouch/WaxData.swift
@@ -9,9 +9,9 @@
 import Foundation
 
 class WaxData {
-    let time:NSTimeInterval, x:Double, y:Double, z:Double
+    let time:NSTimeInterval, x:Float, y:Float, z:Float
     
-    init(time:NSTimeInterval, x:Double, y:Double, z:Double) {
+    init(time:NSTimeInterval, x:Float, y:Float, z:Float) {
         self.time = time
         
         self.x = x
